Add modal for login
