+++ client/package.json
@@ -18,7 +18,7 @@
     "@radix-ui/react-tabs": "^1.0.3",
     "@radix-ui/react-toast": "^1.1.3",
     "@reduxjs/toolkit": "^1.9.5",
-    "daisyui": "^2.51.6",
+    "daisyui": "^3.7.7",
     "dotenv": "^16.0.3",
     "firebase": "^9.21.0",
     "framer-motion": "^10.12.4",
